Replace GenericShader with Magnum's Flat shader.
The goal of this commit is to preserve current visual output while
porting to Magnum's builtin functionality, in order to ensure no
regression happened before we jump on more complex shaders.

There's one feature of the GenericShader that still has to wait for
changes on Magnum side and that's the primitive ID texturing used by PLY
files. The relevant shader code is separated into a new
PrimitiveIDTextured{Shader,Drawable} where it will temporarily live
until Magnum can handle this as well.