GPU-based depth unprojection (#164)
* Add a test and a benchmark for the depth unprojection code.

* Further simplify depth unprojection code.

And make the docs public and nicely formatted. This improves the runtime
by about 20%, however without the branching it would be much much faster.

Co-authored-by: Mandeep Singh Baines <msb@fb.com>

* Implement GPU-based depth unprojection and a depth-only shader.

* Benchmark also branchless depth unprojection for comparison.

* Pass TARGET_HEADLESS to Magnum's CMake when building headless.

So it uses EGL and not GLX for windowless apps.

* Do the far depth patching in a separate loop.

Could have a positive impact on vectorization in the first loop.

* Updated Magnum with WindowlessEglApplication improvements.

Also includes an update of OpenGLTester where it doesn't enable debug
output by default to make the output less spammy.

* Make depth unprojection test comparison fuzzier.

* Try GCC's FMV on the unprojection code (and benchmark).

SSE 4.2, AVX2 and a default fallback for now. I need to figure out what's
the correct way to enable AVX512.

* ci: print CPU info for each job.