Make semantic mesh shader work with WebGL2
The semantic mesh shader relies on gl_PrimitiveID which is not
available in GLES3.0 and hence WebGL2. Instead, create a combined
set of indices.

Tested with 17DRP5sb8fy_semantic.ply:

vertices before: 1522546
vertices after:  1572308

Since the increase is only 3%, the benefits of having a separate
gl_PrimitiveID shader are not large enough to warrant the complexity
of supporting two shaders.

I've only tested one mesh so the overhead could be larger on other
meshes.