Docs: new light theme and various other updates (#209)
* docs: dang, I forgot to clean this up.

* docs: relative paths actually *do* work.

Not sure why I thought they aren't.

* Cross-link to the quaternion package as well.

For some reason I grabbed a totally broken inventory file at first. This
one is less broken but the names get inspected weird so it needs to be
fixed up after.

* Move the build_docs.sh helper into the docs directory.

* Updated m.css submodule.

* docs: capitalize project name.

* docs: move the generated tagfile outside of the doxygen output dir.

Somehow it doesn't generated when it's there. Or something.

* docs: make it possible to supply external m.css.

* docs: switch to a custom light theme.

The resulting CSS is generated by the build.sh script and added to
.gitignore -- changes should be made to theme.css / pygments-pastie.css
instead.

* docs: link to tutorials from the main page.

Similarly to how API does that.

* docs: update magnum bindings doc inventory.

The URL scheme got stabilized there, so all links are different now.

* docs: less confusing names in the navbar.

* docs: make main project URL just a /.

So it's "relocatable", i.e. when running the whole site on localhost,
the navbar links goes to the main site instead of out into the wild.