docs: sync doc theme with website updates, random fixes. (#252)
* docs: sync doc theme with website updates.

This affects habitat-api as well, as that repo looks here for the CSS.

* Updated m.css submodule.

With some minor updates and a crash fix for a corner case in the typing
module.

* docs: due to the new __all__, utils.common got lost from the docs again.

* docs: TODO and bug lists don't belong into public docs.

* docs: a separate public config with GA and improved search.

Adding GA tracking to the docs, but to avoid privacy issues in a local
build, there's now a separate public config. Additionally, for the
public version the search downloads a binary data (which is smaller, but
won't work locally), plus also provides a helpful link to Google search
if nothing is found.