Bundle Basis Universal transcoder and enable the BasisImporter plugin.
https://github.com/BinomialLLC/basis_universal/tree/2f43afcc97d0a5dafdb73b4e24e123cf9687a418

Unfortunately the repo has >200 MB history due to huge files added in
the past, so it's not feasible to add it as a Git submodule. Disabling
the (huge) BC7m6 format and a few others that Magnum has no support for,
the BC7m6 file is empty because the sources include it unconditionally.
Fix in https://github.com/BinomialLLC/basis_universal/pull/91.