Update to Magnum 2019.10 (#334)
* Update to Magnum 2019.10.

To be clear, the v2019.10 tag in all repos. Also updating all Find
modules to latest to make system-magnum builds work.

* Bundle Basis Universal transcoder and enable the BasisImporter plugin.

https://github.com/BinomialLLC/basis_universal/tree/2f43afcc97d0a5dafdb73b4e24e123cf9687a418

Unfortunately the repo has >200 MB history due to huge files added in
the past, so it's not feasible to add it as a Git submodule. Disabling
the (huge) BC7m6 format and a few others that Magnum has no support for,
the BC7m6 file is empty because the sources include it unconditionally.
Fix in https://github.com/BinomialLLC/basis_universal/pull/91.

* assets: setup for Basis target format, compressed image import.

* List a forgotten USE_SYSTEM_GLFW option.

It was used internally, but never exposed to CMake GUI / ccmake.