assets: unconditionally interleave imported glTF mesh data. (#554)
Since not all glTF files have interleaved buffers, the GPU had to fetch
vertex attributes scattered across memory instead of having them nicely
localized. The original MeshData3D-based workflow interleaved the
attributes always during compile(), but since the new workflow aims to
minimize data processing, a slice of the data comes from the file
unchanged. The interleave() API is a no-op in case the mesh is already
interleaved, and it doesn't allocate anything if a r-value is passed so
this should have a strictly positive effect on performance.

This currently also resolves an issue with heavier memory usage, which
is due to the input glTF files being laid out in a way that hits a
suboptimal code path in TinyGltfImporter. A fix on that side will arrive
later.