Exclude test code from coverage reports. (#557)
Tests for C++ and Python code are one third of the total code size
(~1.7kLoC, while the total code size is ~5k), and as it usually is with
tests, those are ran in full with >95% coverage. This however skews the
coverage percentage into suspiciously high numbers, making us think the
codebase is covered better than it actually is.

This is consistent with how coverage is measured for the Magnum
projects, and I'm sure is among best practices in many other projects.