Local planner (#546)
The current greedy follower wasn't designed in a world were sliding was disabled and action noise existed, so it isn't great in those situations. This PR switches it out for a local motion primitive planner that does a considerably better job and also sets things up for action noise (by allowing the user to specify what key should be emitted for each action to take as the planner needs to plan on noiseless actions).

Compared to the old one, there is a bump of ~0.02 SPL when sliding is enabled. When sliding is disabled however, SPL is increased from ~0.85 (with a ~12% failure rate) to ~0.95 (with a ~2% failure rate), which is much better.