doc: don't alias public and local tagfile location.
Yet another case of Doxygen unable to create a directory, however this
time it fails silently. Sigh.

This fails when generating public docs and the build/docs/ directory
doesn't exist. The m.css scripts creates build/docs-public/ in that
case, so tell Doxygen to put everything there, including the tagfile.