Updated Corrade to latest.
This introduces new String and StringView classes, which were not
included in the previous update as those triggered an issue with ar on
Emscripten, and an attempted fix for that issue triggered an issue with
CMake < 3.13, and actually I just spent two days on an issue in
Emscripten that was fixed *by us* in the first place, a year ago.

Additionally fixes an issue with generating Magnum version headers that
was shown as a warning during CMake configure step.