--Make default attributes templates read-only.
Expand default attributes handling to be available to all attributes managers.  Makes object templates specified by physics world attributes or scene attributes, as well as those associated with primitive assets, to be unable to be removed by the user.  Adds a single function and python binding to be able to remove all removable templates in one command.  Adds tests to verify functionality.