--Refactor so that template type is verified to be valid.
Previously, the attributes managers specialized the class template by specifying a smart pointer to an attributes class as the class template type parameter.  This has been changed to just use the attributes class as the template type to specialize the class template.  This allows us to verify that the specified template class is derived from the desired base class (so we can be sure that the objects we wish to manage with a particular manager implement the necessary functions for the attributes managers to work correctly).  If an illegal class is passed now as a template parameter, the code will not compile.