Disabled running version_conda_release on every master commit (#797)
After latest change we noticed that Conda build step was triggered both for tagged commit and for the one merged into master. Based on CircelCI documentation that's the right way to disabled running version_conda_release on every master commit as filters section is union not intersection of branch and tag filters.