-- add source value to scene instance, to represent where scene was made.
If a scene instance is composed in blender, no notion of the COM will exist and any translations will need to be corrected by COM offset.  However, in the future we plan on supporting the saving of existing, composed scenes to scene instance attributes, and the locations of objects will be already corrected for their COM offset.  So this introduces a mechanism to differentiate between the two processes.