--Remake activateCollisionIsland with bitset
std::set.insert can throw an exception, but BulletRigidObject::activateCollisionIsland was being called from the destructor, which cannot allow exceptions.  Instead of bypassing clang-tidy check, using a bitset will never throw an exception.  A potential issue is if we have more than 65536 island ids, this will overflow.  If this is deemed too small, it can easily be expanded.