[Bugfix]: make corrade plugins static (#1480)
Previously our corrade plugins were not statically linked which complicated the rpath resolution. Apparently, this was a bug and can be fixed with a simple cmake change which ensure that corrade is built statically as well.
Now that this change is made, it vastly simplifies the rpath issues we were having before since Magnum and Habitat-Sim are both separately and statically linked.
Fixes a bug where importing magnum would fail if habitat_sim was not imported first (due to shared lib dependencies installed in the habitat_sim folder.