Don't include MetadataMediator in Simulator.h and ResourceManager.h.
Not exactly sure what is it doing, but it's quite heavy. Unfortunately
most of the tests need it to set things up anyway, so it only helps
partially.

Also, the ::ptr typedefs are nice and all, but they mean one has to
include the whole class definition. Which is not necessary in 99% of
cases, so I'm converting T::ptr to the verbose std::shared_ptr<T> where
a forward declaration is enough.