Random "header hygiene" cleanups (#1880)
* Remove unused and wrong Magnum typedefs.

Those have a hardcoded assumption that any mesh is a GL mesh, and that
any material is a Phong material. Which is kinda true now, but it won't
be going forward.

Moreover, they weren't really used anywhere anyway, so there's no point
in having them at all.

* Don't include all of Magnum everywhere for no reason.

The magnum.h header now contains SceneGraph forward declarations
typedefs and only that. It doesn't and shouldn't pull in also GL or
Trade stuff.

* Contain Eigen into a separate header, not the main main header.

It's still pulled into far too many places, but this way it can at least
start getting gradually removed.

* Use Magnum's forward declaration headers, not manual declarations.

* Don't include MetadataMediator in Simulator.h and ResourceManager.h.

Not exactly sure what is it doing, but it's quite heavy. Unfortunately
most of the tests need it to set things up anyway, so it only helps
partially.

Also, the ::ptr typedefs are nice and all, but they mean one has to
include the whole class definition. Which is not necessary in 99% of
cases, so I'm converting T::ptr to the verbose std::shared_ptr<T> where
a forward declaration is enough.

* Remove most includes from ResourceManager.h.

Wow, imagine changing anything in any of those headers. Boom, everything
gets recompiled.

* Stop including ResourceManager from all headers.

* This is not the right way to create a string constant.

The static will *copy* the string into each and every object file that
directly or indirectly includes Asset.h.

* Disable a useless Clang Tidy check.

* Fix definition of the global logging context variable.

It only worked so far because <Magnum/configure.h> was not included in
Logging.cpp and so MAGNUM_BUILD_STATIC_UNIQUE_GLOBALS wasn't defined.
Now it is, and the anonymous namespace is in direct conflict with the
export macro.