Assorted fixes for more robust handling of asset corner cases (#1915)
* Enable shader texture transform only if it's actually textured.

If it's not (for example if a texture failed to load or if the
material data are strange in some way), then the shader construction
would assert with

    Shaders::PhongGL: texture transformation enabled but the shader is not textured

This change *doesn't* fix any root causes of a texture missing or
failing to load, it only makes it not abort anymore.

* Don't crash if BasisImporter isn't present.

* Don't treat files with no default scene as having no scenes at all.

Magnum's GltfImporter used to return 0 if the default scene field wasn't
present, but now it returns -1 to indicate no default scene is
specified, consistently with all other importers. The code mistakenly
relied on the previous behavior.