gfx_batch: change addMeshHierarchy() to bake in the initial transform.
As there's already an API to change the transformation of a hierarchy
right after it's added, this parameter was kinda redundant.

But I need to have a way to set an initial "coordinate frame correction"
transformation and keep it there, forever, which is what this parameter
could do instead -- so now it applies the transform to the immediate
child nodes instead, leaving the top-level transformation as an identity.