gfx_batch: add a more efficient way to get images on the CPU.
I don't intend these to be used that much, nevertheless it's good to not
do stupidly inefficient things such as allocating and then copying a
huge buffer every frame just because "it's fine for a test".

It also allows for easy querying of image subranges, which is a nice
bonus.