gfx/replay: use a virtual interface instead of heavy std::function.
Also putting that back to a single file -- a Player can't really be
used without the implementation class anyway so a separate header
doesn't make sense.

The light setup and semantic ID setting is optional, with the default
implementation being a no-op.

Also changed gfx::replay::GfxReplayNode* to gfx::replay::NodeHandle.
It's shorter and the pointer is a part of the typedef, thus removing a
potential ambiguity about whether one should use a * with this type or
not.