Give up and use a shared_ptr for PlayerImplementation.
It was crashing because the python bindings for
ReplayManager::readKeyframesFromFile() returned a Player but didn't tell
Python the Player depends on the PlayerImplementation stored inside the
ReplayManager. Which would involve some additional bookkeeping in the
bindings, and ... everything else is using a shared_ptr anyway, so
trying to avoid the unnecessary overhead in just a single place of all
is like a drop in the ocean.

Doesn't crash anymore with this change.