--[PBR Update 2 of 3] PBR rewrite-motivated default lighting and IBL fix/reworking (#2119)
* --scale lights only in PBR shader;
Default lights should work well for both PBR and Phong shading. By moving the scaling value to the PBR shader, the lights can now be agnostic to shader.
* --better light balance
The PBR 2.0 shader refactor will remove the hardcoded scaling by a change of units conversion.
* --rebalance IBL mix
* --make default lights camera-relative (temporarily)
* --modify replay test to account for added default light.
* --change case of ibl flag
* --enable viewer.py ibl flag
* --better default lights
* --fix simConfig comparator to cover IBL flag
* --move IBL instantiation out of RM constructor
Having the IBL env map loaded in the ResourceManager constructor was failing since there usually wasn't a GL context built yet, which was needed to instantiate the shader.
This is just a stop-gap measure so that IBL can be invoked successfully.  Much engineering needs to be done for IBL to be fully ready for prime time.
* --minor default light adjustment
* --updated test images for new lighting
NOTE :  Some of these images will need to be updated again with the full PBR fragment shader rewrite.
* --add hdr image;
* --removed camera-relative lights from default light setup;

* --tweaked and removed unnecessary downward light