--handle empty stage instance better.
The default scene instance attributes will have all fields empty, so a stage_instance can be created with a template name of NONE, corresponding to the NONE stage.

However, if a non-default scene instance configuration is loaded and it is lacking the stage_instance field, that is indicative of a dataset error (There always must be a stage instance specified for every scene instance) and so exit the program so that it can be fixed.