--[BE Week]More informative and appropriate data/metadata load and processing messages (#2144)
* --improve MM data load messages; Some errors->ESP_Check
* --improvements to debug messages in RM
Some demotions to verbose;
* --clarify what "Aborting" means when recoverable data errors are found.
We often use 'Aborting' in failed assertion messages, usually implying a core dump. This commit clarifies when processes fail due to data errors, not so grievously that an assertion/std::abort is called for. These failures are often caught downstream by ESP_CHECK.
* --handle empty stage instance better.
The default scene instance attributes will have all fields empty, so a stage_instance can be created with a template name of NONE, corresponding to the NONE stage.

However, if a non-default scene instance configuration is loaded and it is lacking the stage_instance field, that is indicative of a dataset error (There always must be a stage instance specified for every scene instance) and so exit the program so that it can be fixed.
* --slight refactor of AttributesManagers' JSON handling.
Function used to load, but now it just sets the attributes based on already loaded JSON.