--PBR/IBL Configuration and customization (#2155)
* --initial assignment of gamma and exposure as uniforms
* --pbr/ibl shader attributes and mgr; default shader config
* --add PbrShaderAttributesManager to MM
* --handle non-reset material spec; remove class ref to material
want to preserve the flags when a material is only being modified so that a new shader is not inadvertently created.
Material class ref removed since we're using the material cache instead.
* --Remove out-of-date 'hacky scaling' for default phong materials and phong->PBR materials
Equalizing brightness between PBR and Phong will now take place via PbrConfig settings.
* --Drawable Configuration
pass a configuration structure to the drawables on creation that will be created beforehand and hold appropriate references to various configuration quantities based on the type of drawable.
* --initial stubs for PbrShaderAttribute handling
* --initial SceneDataset/SceneInstance config proc for PbrShaderAttributes
* --add formatted output stream for Configurations
* --set PbrAttributes properly before drawable call
* --move image loading and texture config to RM; Cache loaded IBL textures
* --clarify move construction of mat cache on reset
* --enable turning on/off IBL across all shader configs at once.
* --fix missing context issue.
IBL assets should only be loaded and processed when a context exists and when MM is in its final state during load.
* --missed hacky scaling instance
No need to scale phong colors any more to handle over-bright lights that were needed for the old PBR shader.  Any PBR-driven scaling will be accomplished via PbrShaderAttributes.direct_light_intensity field.
* --update test images to reflect non-scaled phong.
* --change pbrIBLHelpers storage to be a map
The pbrIBLHelpers hold access to the bLUT and the env maps built from the loaded environment map asset. This now stores them in a map that is keyed by the truncated names of the bLUT and envMap source assets, to facilitate access and prevent rebuilding the same assets.
* --initial mappings of configs into pbr drawables.
* --second step - direct and indirect flags
* --third step - expand config support.
* --handle IBL tonemap variable; intensity scale
* --set Burley vs Lambertian diffuse direct
Lambertian tends to look washed out, particularly on bright surfaces, while Burley looks much better.
* --remove unused shader config setting, deprecate IBL directive
PbrShaderAttributes needs to be governed by Scene datasets and MetadataMediator, and IBL status is based on config settings now.
* --add PBR Shader Attributes and PbrShaderAttributesManager tests
* --complete bindings coverage for Pbr shader attributes and managers
* --add env maps and reorganize compiled resources.
--Remove old lythwood image
--Add 4 new 1k hdr envmaps
* --region-based PbrShaderAttrib queries
* --add config prototypes for additional env maps
* --Improve gamma and exposure mapping
* --fix issue with IBL results always mapping to sRGB
--also provide more expansive srgb <--> linear control
--The default IBL image has also been changed to one that is not so gold/orange.
* --modify PbrIBLHelper construction
--If a user modifies or adds a new PbrShaderConfig with new IBL assets, while RM exists without re-running sim::reconfigure and tries to instantiate an object using this new PbrShaderConfig, those new assets will not have been loaded yet. This change makes it so that if they haven't been, they will be loaded and cached.