--add restrictions to possible registration
To be valid, an Articulated Object attributes must reference an existing, findable URDF source file and, if it specifies 'Skin' render mode, it must also reference an existing, findalble skin render asset. If these conditions are not met, registration will fail.  These restrictions are analogous to the restrictions we place on Object and Stage attributes with regard to their render assets.