--populate default urdf_filepath in config with reasonable default value
The default value will be the same as the config's name but with the extension '.urdf' instead of '.ao_config.json'. This means that for this to work - for out-of-date ao_config.json files to still work properly, their .urdf counterpart must have the same name as they do (minus extension) and reside in the same directory that they do.  This is analogous to how ao configs are currently being loaded, so this should be expected.