--Articulated Object Attributes/Managers support (#2178)
* --initial commit of articulated object attributes,AO Attributes Manager,AO mgrs to MM and SceneDatasetAttrs; removal of hand-rigged AO load code in scene dataset load; bindings;initial refs to AO attr/mgr in RM and Sim
* --add AO attrs/managers to python module list.
* --set filepath for render asset if listed path is not found.
As with all JSON configs, the asset filepath is expected to be relative to the config's location on disk.
* --add support for Articulated Object Render Mode enum
--This specifies how to render the AO - using its skin, link visualization assets, both or neither.
* --update AttributesConfigsTest for ArticulatedObjectAttributes
* --add bindings for URDF filename; add mgrs tests for AOAttributesManager
* --add restrictions to possible registration
--We want to fail on load of attributes configs if certain criteria are not met. To be valid, an Articulated Object attributes must reference an existing, findable URDF source file and, if it specifies 'Skin' render mode, it must also reference an existing, findalble skin render asset. If these conditions are not met, registration will fail.  These restrictions are analogous to the restrictions we place on Object and Stage attributes with regard to their render assets.
* --Update the AttributesManager test to properly test for required files
Object, Stage, and Articulated Object attributes can be built from the file handle of a non-config file. For Object and Stage attributes, they can be built from the name of an existing render asset, while for Articulated Object attributes they can be built from the name of an existing URDF file. The AttributesManagers for these 3 configs have been updated to test for proper building and registration of default attributes from these required asset filenames, including verifying that those assets exist before the built attributes can be registered.
* --MM : verify that a new SceneInstanceAttributes is created properly
--Shouldn't ever happen, but safety first.
* --populate default urdf_filepath in config with reasonable default value
The default value will be the same as the config's name but with the extension '.urdf' instead of '.ao_config.json'. This means that for this to work - for out-of-date ao_config.json files to still work properly, their .urdf counterpart must have the same name as they do (minus extension) and reside in the same directory that they do.  This is analogous to how ao configs are currently being loaded, so this should be expected.
* --add enums covering override-able state values, and attributes support
* --add enum map query function to attributes managers
Will query a json doc for a string key, if present will validate the value against the passed map keyed by strings with enum class values. The only constraint on the enum class is that it has an unspecified value (ideally first, mapped to ID_UNDEFINED). This way these map functions can be used for both config values and instance config values, where 'unspecified' would mean use the underlying config value.
* --add AO scene instance support for enum-backed values
Retain boolean "fixed base" query for now, but this will be removed in favor of only having the enum-backed value.
* --move URDF parser to metadata subsystem
* --build Articulated Object using Attributes.
* --repo add-ignore reference to replica_cad dir required for tests
* --remove "debug" and "primitive" refs to AO visual shapes and rendering
* --enable physics in ReplicaCAD test so that URDFs will be loaded