--[BUGFIX] -Ellipses in paths (#2218)
* --filter paths of ellipses and their parent, referencing directories
If an OS link spans the ellipsis, certain functions, like glob, get lost when attempting to navigate the ellipsis.
* --add test
* --fix for glob call
--glob doesn't know how to back-step across an os link, so it gets lost if it encounters one spanned by an ellipsis.
* --rename to normalizePath for consistency with other similar functions.
* --tailor the functionality only to specific use case
In an effort to limit the impact of this change, it has been narrowed to only affect paths prefixed with an ellipse in the scene dataset config. Since the scene dataset config is expected to exist at the root of the dataset, and since the scene dataset may be accessed via a link, having an ellipses from the dataset root back-pedals across a link. Some functions, like glob(), don't handle this properly and get lost. This commit addresses only these cases.