--Region query tools (#2336)
* --add region poly loop area calc; add weighted point containment check
To support nested regions, this check will provide weighted results, with the highest weighted region being the one with the smallest area.
* --cleanup/organize semantic functions
* --add weighted region checking for a point, to account for nested regions
The weighting is derived from the region area, where the smallest region a point is found in has the highest weighting.  If a point only resides in a single region, its weighting is 1.
* --give every containing region an equal vote (in the case of nested regions)
For nested regions, every containing region will get an equal vote for the multiple points. This means that a set of points that are mostly in the inner region of a nested region will still appear to be 'more' in the outer region that contains all the points.