--provide mechanism for correcting improper marker load locations due to inappropriate COM correction.
Due to a quirk with how object attributes are read and saved, and how a specified COM vector vs an unspecified COM vector in the configuration json is treated by the engine when determining the object origin location, points that were saved with the original configs (which had no COM specified and so therefore relied on rigid object COMs being determined based on bbox center) would be saved in configs that would get a 0,0,0 COM saved as well, inappropriately. Upon subsequent load, this 0,0,0 would be used to determine the object's origin, instead of the bbox COM that was used initially, and the marker points would be displaced from where they belonged. This PR enables the reversal of this displacement.

PR#2382 will correct the inappropriate save of COM location, and a subsequent PR will include the internal bindings to the uncorrected COM location that are added to this commit as well.