--Add Configuration State Flags - Part 1 (#2396)
* --set ConfigValue to have 8 byte type and flags member
* --add support for "isDefaultVal"
If ConfigVal values are set programmatically to defaults, the process should use "init" instead of "set". This will make sure the isDefaultVal flag is set to true, and consequently not writing those values to file when requested to.
Values set by the existing "set" methods will always be written to file.
* --change 'set' to 'init' in constructors for attributes
Nearly all the values set in the constructors of the various attributes are defaults that are expected to be overwritten only if necessary.  Writing these to file is redundant and can, in certain circumstances,  cause written JSONs to be in an inappropriate state.
* --Make sure "is_instance_visible" stays an int
* --clarify needlessly arcane data size check.
* --add flag for a hidden internal value
A hidden value would be a pertinent configuration value that is set programmatically and only used internally. These values should not be written to file or exposed to the user except perhaps for debugging.
* --check before destructing dest *ptr that it isn't the same as src *ptr
Not sure if this could even happen (a situation could arise where two different instances could have their _data arrays both be pointers to the same memory location, but just in case.
* --Only write ConfigValues to file if they are not hidden or defaults
* --add hidden value setters
* --simplify orientation frame handles
* --setup hidden fields
Hidden fields are set and used internally and will generally not be written to file or exposed to the user except in a debugging capacity. Their labels in Configurations should be camel-case with 2 leading underscores, to help differentiate them from non-hidden fields.
* --another hidden value
'Force_flat_shading should be considered a hidden value, only used internally.
* --cleanup; fix erroneous shouldWrite check; set frame init in obj/stage
* --add "isTranslated" flag to denote enum-backed values
This means a config value is translated by consuming code to some other representation, (i.e. an enum) before it is consumed. These values are stored internally and in files as strings, but require the implementation of custom read functionality to make sure the value read is within the acceptable bounds supported by the program.
* --initialize Mn::Vector3s with floats
* --Config store AssetType like other enums; Handle init'ing enum vals
* --improve AssetType attributes internal handling
* --Internal tool to rebuild an abs path to be relative to another abs path.