--enable spot to disengage and re-engage the navmesh (i.e. like no_clip)
Turning no_clip on disengages the navmesh. Turning noclip off looks for the closest point on the largest island, than the closest point on any part of the navmesh first, and if one is found, then it snaps to it, otherwise it stays in noclip mode, gives a message to get closer to the navmesh.