--handle if the 'markers' subconfig does not exist in a MarkerSet
It is possible to have a JSON marker set object associated with a link and task that lists no points. Due to the markersets being managed generically as nested configurations, with access accomplished via static casting, it is then possible to access a MarkerSet object that may not have a "markers" subconfig. This is a reasonable state, and should result in an assertion when attempting to query the MarkerSet.